+++ src/cli.rs
@@ -1,4 +1,4 @@
-use clap::{Parser, Subcommand, builder::RangedU64ValueParser};
+use clap::{builder::RangedU64ValueParser, Parser, Subcommand};
 
 #[derive(Parser)]
 #[command(